requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.6
debugpy==1.8.17
decorator==4.4.2
defusedxml==0.7.1
dm-sonnet==2.0.2
dm-tree==0.1.9
duckdb==1.4.1
exceptiongroup==1.1.3
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graph_nets==1.1.0
grpcio==1.74.0
h11==0.16.0
h5py==3.15.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==6.25.2
ipython==8.22.2
ipywidgets==8.1.7
isoduration==20.11.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.15
keras==2.12.0
kiwisolver==1.4.9
lark==1.3.0
libclang==18.1.1
lightning-utilities==0.15.2
llvmlite==0.45.1
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mistune==3.1.4
ml_dtypes==0.5.3
mplsoccer==1.5.1
mpmath==1.3.0
multidict==6.7.0
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nimfa==1.4.0
notebook_shim==0.2.4
numba==0.62.1
numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
polars==1.35.0
polars-runtime-32==1.35.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.0
pure_eval==0.2.3
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-pptx==1.0.2
pytorch-lightning==2.5.5
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
statsmodels==0.14.5
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.20.1
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.3.0
torch==2.9.0
torch-geometric==2.6.1
torch_geometric_temporal==0.56.2
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==2.9.0+cpu
torchmetrics==1.8.2
torchvision==0.24.0+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.14.2
xgboost==3.1.1
xlsxwriter==3.2.9
yarl==1.22.0
